adds prelim scopes with filterable module; updates the square mile param to be an integer

+++ db/migrate/20170721165425_parks_table.rb
@@ -11,7 +11,7 @@ class ParksTable < ActiveRecord::Migration[5.1]
       t.string :closure
       t.string :amenities
       t.string :activities
-      t.decimal :sq_mi
+      t.integer :sq_mi
 
       t.timestamps
     end
